+++ requirements.txt
@@ -1,0 +1,18 @@
+tensorflow-datasets==4.4.0
+tensorflow-estimator==2.4.0
+tensorflow-metadata==1.2.0
+termcolor==1.1.0
+threadpoolctl==2.1.0
+tifffile==2021.8.30
+tokenizers==0.9.3
+torch==1.7.1
+tqdm==4.62.3
+traitlets==5.1.0
+transformers==3.5.1
+tweepy==3.10.0
+typing-extensions==3.7.4.3
+urllib3==1.26.7
+wcwidth==0.2.5
+Werkzeug==1.0.1
+wrapt==1.12.1
+zipp==3.5.0
